Update models.py
For a large N variables the matrix `np.zeros((N, tau_max +1, N, N)` scales quickly in RAM requirements. For those cases where there is not enough RAM, we try to lower the precision to `float16` which is of half precision but requires a quarter of the RAM. We raise a warning for the user.