Update parameter of pc_stable for pcmci
In the documentation it is stated that default pc_alpha is a list (default: [0.05, 0.1, 0.2, ..., 0.5]).

On the other hand, the default value is 0.2 which is not converted to this list. By setting pc_alpha to None, the code converts it to the stated list.

Note: another option to solve this discrepancy is to modify the documentation, but I find it more interesting to have the optimization of pc_alpha as the default case.